Capture configuration state before initialization
To allow parallel concurrent subscriptions we need to capture all
operator configuration state before initialization. Here we leverage
MemberwiseClone to do a shallow copy of the entire object state.

This will allow for derived implementations to keep overriding the
Configure method and accessing instance state, since the
operator will ensure it is captured. Other designs are possible but
would introduce breaking changes.

Implementors should be aware not to use mutable reference classes as
configuration state, as these will not be correctly copied.